app/new_requirements.txt
Authlib==1.3.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
graphdatascience==1.9
idna==3.7
multimethod==1.10
neo4j==5.13.0
numpy==1.24.4
pandas==2.0.3
pyarrow==14.0.2
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
textdistance==4.6.1
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
validators==0.28.1
weaviate-client==3.24.2
